refactor(pxe): replace Oracle class with Proxy-based buildACIRCallback (#23915)
