+++ noir-projects/aztec-nr/aztec/src/macros/internals_functions_generation/external/public.nr
@@ -112,8 +112,12 @@ pub(crate) comptime fn generate_public_external(f: FunctionDefinition) -> Quoted
 
     let (assert_initializer, mark_as_initialized) = if is_fn_initializer(f) {
         (
-            quote { aztec::macros::functions::initialization_utils::assert_initialization_matches_address_preimage_public(self.context); },
-            quote { aztec::macros::functions::initialization_utils::mark_as_initialized_from_public_initializer(self.context); },
+            quote {
+                aztec::macros::functions::initialization_utils::assert_initialization_matches_address_preimage_public(self.context);
+            },
+            quote {
+                aztec::macros::functions::initialization_utils::mark_as_initialized_from_public_initializer(self.context);
+            },
         )
     } else {
         (quote {}, quote {})
