+++ noir-projects/noir-contracts/contracts/test/storage_proof_test_contract/src/storage_proofs/branch_hash.nr
@@ -19,7 +19,7 @@ fn compute_rlp_length(node: Node) -> u16 {
                 if node.row_exist[$i] {
                     len += 33;
                 } else {
-                    len +=1;
+                    len += 1;
                 }
             }
         },
@@ -51,11 +51,12 @@ fn rlp_encode_to_sink<S: BytesSink>(node: Node, sink: &mut S) {
         for i in 0..16 {
             let mut inner_iterations: [Quoted] = @[];
             for j in 0..4 {
-                inner_iterations = inner_iterations.push_back(
-                    quote {
-                        sink.add_u64(node.rows[$i][$j]);
-                    },
-                );
+                inner_iterations = inner_iterations
+                    .push_back(
+                        quote {
+                            sink.add_u64(node.rows[$i][$j]);
+                        },
+                    );
             }
             let inner_code = inner_iterations.join(quote {});
             outer_iterations = outer_iterations.push_back(
