+++ yarn-project/pxe/src/contract_function_simulator/oracle/index.ts
@@ -1,16 +1,2 @@
-import type { Oracle } from './oracle.js';
-
-export * from './oracle.js';
+export * from './acir_callback.js';
 export * from './interfaces.js';
-
-/**
- * A conditional type that takes a type `T` and returns a union of its method names.
- */
-type MethodNames<T> = {
-  [K in keyof T]: T[K] extends (...args: any[]) => any ? K : never;
-}[keyof T];
-
-/**
- * Available oracle function names.
- */
-export type ORACLE_NAMES = MethodNames<Oracle>;
