+++ yarn-project/pxe/src/contract_function_simulator/oracle/interfaces.ts
@@ -1,5 +1,4 @@
-import type { HandlersForPrefix } from './oracle_registry.js';
-import type { ORACLE_REGISTRY } from './oracle_registry.js';
+import type { HandlersForPrefix, ORACLE_REGISTRY } from './oracle_registry.js';
 
 /**
  * Miscellaneous oracle methods, not very Aztec-specific and expected to be available in all scenarios in which aztec-nr
