+++ yarn-project/txe/src/oracle/txe_oracle_top_level_context.ts
@@ -29,10 +29,10 @@ import {
   ExecutionTaggingIndexCache,
   HashedValuesCache,
   type IMiscOracle,
-  Oracle,
   PrivateExecutionOracle,
   TransientArrayService,
   UtilityExecutionOracle,
+  buildACIRCallback,
   executePrivateFunction,
   generateSimulatedProvingResult,
 } from '@aztec/pxe/simulator';
@@ -857,7 +857,7 @@ export class TXEOracleTopLevelContext implements IMiscOracle, ITxeExecutionOracl
         transientArrayService: new TransientArrayService(),
       });
       const acirExecutionResult = await simulator
-        .executeUserCircuit(toACVMWitness(0, call.args), entryPointArtifact, new Oracle(oracle).toACIRCallback())
+        .executeUserCircuit(toACVMWitness(0, call.args), entryPointArtifact, buildACIRCallback(oracle))
         .catch((err: Error) => {
           err.message = resolveAssertionMessageFromError(err, entryPointArtifact);
           throw new ExecutionError(
